Revert "Bug 2042256 - Adjust fuzz for Skia m149. r=gfx-reviewers,media-playback-reviewers,nical,aosmond" for causing xpc failures at test_imgtools.js
This reverts commit 9f778821407053011a05ddc785f678fd27ace0bf.

Revert "Bug 2042256 - Update mozbuilds for Skia m149. r=gfx-reviewers,nical"

This reverts commit 674dfdde151ddaeae487eb5735d4a320631584b7.

Revert "Bug 2042256 - Fix Moz2D for Skia m149. r=gfx-reviewers,nical"

This reverts commit c9d727ae880e00e97e291f346db252d7856c2193.

Revert "Bug 2042256 - Update Skia source files for Skia m149. r=gfx-reviewers,nical"

This reverts commit dc99955deb8172c40e307661a01b2407bd95ff23.

UltraBlame original commit: 3236899b26fbee3e8b3d5d6ff8e37aa4cc4d43a4

+++ gfx/skia/skia/src/sksl/ir/SkSLConstructor.h
@@ -102,11 +102,11 @@ public:
     }
 
     SkSpan<std::unique_ptr<Expression>> argumentSpan() final {
-        return {&fArguments.front(), (size_t)fArguments.size()};
+        return {&fArguments.front(), fArguments.size()};
     }
 
     SkSpan<const std::unique_ptr<Expression>> argumentSpan() const final {
-        return {&fArguments.front(), (size_t)fArguments.size()};
+        return {&fArguments.front(), fArguments.size()};
     }
 
 private:
